Partial liquid turret implementation / Status effects

+++ core/src/io/anuke/mindustry/entities/units/BaseUnit.java
@@ -74,7 +74,7 @@ public class BaseUnit extends Unit {
 	}
 
 	@Override
-	public void damage(int amount){
+	public void damage(float amount){
 		super.damage(amount);
 		hitTime = hitDuration;
 	}
